Implement database migration setup and update user model
A setup for database migration using Flyway is added in this commit. The user model is updated to make some fields nullable and add new fields such as username and status. The gradle build file is also updated to include dependency for Flyway. This will lead to better database versioning and easier schema changes in the future.